chore: rename svelte component to be consistent with elm component
