+++ app/src/lib/ui/Diagnostic/DiagnosticEdit.svelte
