+++ app/src/routes/(auth)/pro/carnet/[uuid]/diagnostic/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-	import EditDiagnostic from '$lib/ui/Diagnostic/EditDiagnostic.svelte';
+	import DiagnosticEdit from '$lib/ui/Diagnostic/DiagnosticEdit.svelte';
 	import type { PageData } from './$types';
 
 	export let data: PageData;
 </script>
 
-<EditDiagnostic notebookId={data.notebookId} />
+<DiagnosticEdit notebookId={data.notebookId} />
